refactor: reorganize mcp/ into subdirectory with folder segmentation
Move flat prefixed files into extensions/mcp/ directory:
  mcp-types.ts          → mcp/types.ts
  mcp-tool-naming.ts    → mcp/tool-naming.ts
  mcp-transport-stdio.ts → mcp/transport-stdio.ts
  mcp-transport-http.ts  → mcp/transport-http.ts
  mcp-registry.ts        → mcp/registry.ts

Add mcp/index.ts as the folder barrel (re-exports all 5 modules).
Reduce extensions/mcp.ts to a 1-line re-export of mcp/index.ts.

Update all internal imports to use correct relative paths from
within the mcp/ subdirectory (../../config, ../mcp-oauth, ./types, etc.).

No logic changes. All 23 tests pass.
