+++ packages/cli/src/extensions/mcp/tool-naming.ts
