+++ packages/cli/src/extensions/mcp/types.ts
